Changed int phi and int theta to double for all.

+++ Production/Src/SadCL/SadCL_Main/MissileLauncher/IMissileLauncher.cs
@@ -4,7 +4,7 @@ namespace SadCL.MissileLauncher
     public interface IMissileLauncher
     {
         void fire();
-        void moveBy(int phi, int theta);
+        void moveBy(double phi, double theta);
         void reload();
         void status();
         void reset();
